+++ er3t/rtm/mca/mca_sfc.py
@@ -85,7 +85,7 @@ class mca_sfc_2d:
         self.nml['Sfc_nyb'] = copy.deepcopy(self.sfc.data['ny'])
 
 
-        if ('lambertian' in self.sfc.data['alb']['name'].lower()) and (np.squeeze(self.sfc.data['alb']['data']).ndim == 2):
+        if ('lambertian' in self.sfc.data['sfc']['name'].lower()) and (np.squeeze(self.sfc.data['sfc']['data']).ndim == 2):
 
             sfc_tmps = np.zeros((self.sfc.Nx, self.sfc.Ny), dtype=np.float64)
             self.nml['Sfc_tmps2d'] = dict(data=sfc_tmps, name='Temperature anomalies', units='K')    # temperature anomaly
@@ -94,13 +94,13 @@ class mca_sfc_2d:
             self.nml['Sfc_jsfc2d'] = dict(data=sfc_jsfc, name='Surface distribution type', units='N/A')
 
             sfc_psfc = np.zeros((self.sfc.Nx, self.sfc.Ny, 5), dtype=np.float64)
-            sfc_alb  = np.squeeze(self.sfc.data['alb']['data'])
+            sfc_alb  = np.squeeze(self.sfc.data['sfc']['data'])
             sfc_alb[sfc_alb<0.0] = 0.0
             sfc_alb[sfc_alb>1.0] = 1.0
             sfc_psfc[:, :, 0] = sfc_alb[:, :]
             self.nml['Sfc_psfc2d'] = dict(data=sfc_psfc, name='Surface distribution parameters', units='N/A')
 
-        elif ('brdf-lsrt' in self.sfc.data['alb']['name'].lower()) or (self.sfc.data['alb']['data'].shape[-1] == 3):
+        elif ('brdf-lsrt' in self.sfc.data['sfc']['name'].lower()) or (self.sfc.data['sfc']['data'].shape[-1] == 3):
 
             sfc_tmps = np.zeros((self.sfc.Nx, self.sfc.Ny), dtype=np.float64)
             self.nml['Sfc_tmps2d'] = dict(data=sfc_tmps, name='Temperature anomalies', units='K')    # temperature anomaly
@@ -110,12 +110,12 @@ class mca_sfc_2d:
             self.nml['Sfc_jsfc2d'] = dict(data=sfc_jsfc, name='Surface distribution type', units='N/A')
 
             sfc_psfc          = np.zeros((self.sfc.Nx, self.sfc.Ny, 5), dtype=np.float64)
-            sfc_psfc[:, :, 0] = self.sfc.data['alb']['data'][:, :, 0]
-            sfc_psfc[:, :, 1] = self.sfc.data['alb']['data'][:, :, 1]
-            sfc_psfc[:, :, 2] = self.sfc.data['alb']['data'][:, :, 2]
+            sfc_psfc[:, :, 0] = self.sfc.data['sfc']['data'][:, :, 0]
+            sfc_psfc[:, :, 1] = self.sfc.data['sfc']['data'][:, :, 1]
+            sfc_psfc[:, :, 2] = self.sfc.data['sfc']['data'][:, :, 2]
             self.nml['Sfc_psfc2d'] = dict(data=sfc_psfc, name='Surface distribution parameters', units='N/A')
 
-        elif ('cox-munk' in self.sfc.data['alb']['name'].lower()) or (self.sfc.data['alb']['data'].shape[-1] == 5):
+        elif ('cox-munk' in self.sfc.data['sfc']['name'].lower()) or (self.sfc.data['sfc']['data'].shape[-1] == 5):
 
             sfc_tmps = np.zeros((self.sfc.Nx, self.sfc.Ny), dtype=np.float64)
             self.nml['Sfc_tmps2d'] = dict(data=sfc_tmps, name='Temperature anomalies', units='K')    # temperature anomaly
@@ -124,7 +124,7 @@ class mca_sfc_2d:
             sfc_jsfc[...] = 2 # DSM model
             self.nml['Sfc_jsfc2d'] = dict(data=sfc_jsfc, name='Surface distribution type', units='N/A')
 
-            self.nml['Sfc_psfc2d'] = dict(data=self.sfc.data['alb']['data'], name='Surface distribution parameters', units='N/A')
+            self.nml['Sfc_psfc2d'] = dict(data=self.sfc.data['sfc']['data'], name='Surface distribution parameters', units='N/A')
 
         else:
